backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
PyYAML==6.0.2

beautifulsoup4==4.12.3
pypdf==4.3.1

sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1

openai==1.40.6
